When a read() finds no record, it should not raise an exception.
This is a change of behaviour, but moves it in line with the documentation and
all the other code. Previously a missing record would immediately raise a
CantContinue. What happens now is:

* The interface returns an empty Octothorpe.
* An alert of type :error is raised, and @model_status becomes :error
* map_to_model is not called.

If you still want an exception, you should call my_model.read.or_die, of
course.

Note that update() and delete() *do* still automatically raise a CantContinue.
The expectation is that you should already have called read at that point, so
if the record is now missing, something has gone rather wrong.