Release 1.02: Fix a bug in ConnectionPool.
No interface actually releases connections right now, so we were, it turns out, reliant on whatever
process that was spawning threads to do so not excessively.  This was actually happening for me up
to now, even though it's clearly wrong in hindsight.

Even worse: When it didn't work it just hung your application.

New behaviour: if you do not specify a connection pool timeout -- which remains the default -- then
when you reach the maximum number of connections and need a new one, we release the oldest and give
you that.

This is just a fix for that.  I owe Pod4 a major update, but at the moment I'm not in a position to
test for SQL Server or for jRuby.  All the other tests pass.