Deal with the issue of closing DB connections.
In the past this did not appear to be a problem.  I think perhaps that my PostgreSQL was timing out
connections -- it can be set to do that -- but that's hard to prove.

In any case, I'm now observing that connections just hang around until we eventually run out; so it
becomes vital that when we release them in the Connection Pool, we get the interface to tell the
database to drop the connection.

There is a through-test to a PostgreSQL database that demonstrates we are doing this by actually
asking the database about active connections.