Allow a connection pool to be for all interfaces.
This is, FFS, clearly what I was thinking when I designed it, and even said so in the README.  But
since it was based on the earlier Connection class, which for some reason requires that you pass an
interface class to it and then validates incoming interface objects against that -- you could not.

When DB connections are not timing out this is a problem.  What you actually have is a separate
connection pool for each interface, which is a bit too generous.